+++ editor/src/messages/tool/tool_messages/text_tool.rs
@@ -212,7 +212,7 @@ enum TextToolFsmState {
 pub struct EditingText {
 	text: String,
 	font: Font,
-	font_size: f64,
+	font_size: f32,
 	color: Option<Color>,
 	transform: DAffine2,
 }
@@ -235,7 +235,7 @@ impl TextToolData {
 			responses.add(FrontendMessage::DisplayEditableTextbox {
 				text: editing_text.text.clone(),
 				line_width: None,
-				font_size: editing_text.font_size,
+				font_size: editing_text.font_size as f64,
 				color: editing_text.color.unwrap_or(Color::BLACK),
 				url: font_cache.get_preview_url(&editing_text.font).cloned().unwrap_or_default(),
 				transform: editing_text.transform.to_cols_array(),
@@ -326,7 +326,7 @@ impl TextToolData {
 	fn get_bounds(&self, text: &str, font_cache: &FontCache) -> Option<[DVec2; 2]> {
 		let editing_text = self.editing_text.as_ref()?;
 		let buzz_face = font_cache.get(&editing_text.font).map(|data| load_face(data));
-		let subpaths = graphene_core::text::to_path(text, buzz_face, editing_text.font_size, None);
+		let subpaths = graphene_core::text::to_path(text, buzz_face, editing_text.font_size as f64, None);
 		let bounds = subpaths.iter().filter_map(|subpath| subpath.bounding_box());
 		let combined_bounds = bounds.reduce(|a, b| [a[0].min(b[0]), a[1].max(b[1])]).unwrap_or_default();
 		Some(combined_bounds)
@@ -381,7 +381,7 @@ impl Fsm for TextToolFsmState {
 				});
 				if let Some(editing_text) = tool_data.editing_text.as_ref() {
 					let buzz_face = font_cache.get(&editing_text.font).map(|data| load_face(data));
-					let far = graphene_core::text::bounding_box(&tool_data.new_text, buzz_face, editing_text.font_size, None);
+					let far = graphene_core::text::bounding_box(&tool_data.new_text, buzz_face, editing_text.font_size as f64, None);
 					if far.x != 0. && far.y != 0. {
 						let quad = Quad::from_box([DVec2::ZERO, far]);
 						let transformed_quad = document.metadata().transform_to_viewport(tool_data.layer) * quad;
@@ -397,7 +397,7 @@ impl Fsm for TextToolFsmState {
 						continue;
 					};
 					let buzz_face = font_cache.get(font).map(|data| load_face(data));
-					let far = graphene_core::text::bounding_box(text, buzz_face, font_size, None);
+					let far = graphene_core::text::bounding_box(text, buzz_face, font_size as f64, None);
 					let quad = Quad::from_box([DVec2::ZERO, far]);
 					let multiplied = document.metadata().transform_to_viewport(layer) * quad;
 					overlay_context.quad(multiplied);
@@ -409,7 +409,7 @@ impl Fsm for TextToolFsmState {
 				tool_data.editing_text = Some(EditingText {
 					text: String::new(),
 					transform: DAffine2::from_translation(input.mouse.position),
-					font_size: tool_options.font_size as f64,
+					font_size: tool_options.font_size as f32,
 					font: Font::new(tool_options.font_name.clone(), tool_options.font_style.clone()),
 					color: tool_options.fill.active_color(),
 				});
